+++ crates/objc2/src/topics/about_generated/CHANGELOG.md
@@ -45,6 +45,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - `IntentsUI` / `objc2-intents-ui`.
   - `IOBluetooth` / `objc2-io-bluetooth`.
   - `IOBluetoothUI` / `objc2-io-bluetooth-ui`.
+  - `iTunesLibrary` / `objc2-itunes-library`.
   - `JavaScriptCore` / `objc2-java-script-core`.
   - `LatentSemanticMapping` / `objc2-latent-semantic-mapping`.
   - `MediaAccessibility` / `objc2-media-accessibility`.